% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{print.SGCBResults}
\alias{print.SGCBResults}
\title{Print SGCBResults}
\usage{
\method{print}{SGCBResults}(x, ...)
}
\arguments{
\item{x}{SGCBResults object}

\item{...}{Additional arguments}
}
\description{
Print SGCBResults
}
