% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SGCBConfig-class}
\alias{SGCBConfig-class}
\title{SGCBConfig class}
\description{
Stores SGCB analysis configuration parameters.
}
\section{Slots}{

\describe{
\item{\code{bootB}}{Number of bootstrap replicates used when \code{bootstrap = TRUE}}

\item{\code{nCores}}{Number of parallel cores}

\item{\code{seed}}{Random seed for reproducibility}
}}

