% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{sgcbDE-deprecated}
\alias{sgcbDE_fast}
\alias{sgcbDE_infogeom}
\title{Deprecated SGCB Functions}
\usage{
sgcbDE_fast(counts, group, ...)

sgcbDE_infogeom(counts, group, ...)
}
\arguments{
\item{counts}{Count matrix (genes x samples)}

\item{group}{Group vector}

\item{...}{Passed to \code{sgcbDE}}
}
\description{
These functions are deprecated. Use \code{\link{sgcbDE}} instead.
}
